Ajout et modification devaluation,login,register,affichage des détails:projet,dev

+++ src/main/java/ma/ac/uir/synthesedemo/dao/ProjetRepository.java
@@ -4,5 +4,5 @@ package ma.ac.uir.synthesedemo.dao;
 import ma.ac.uir.synthesedemo.entity.Projets;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjetRepository extends JpaRepository <Projets,Integer> {
+public interface ProjetRepository extends JpaRepository <Projets,Long> {
 }
